Support cases when response of xstreamInfo does not contain entries.
When redis stream does not contain any entry fields 'first-entry'
and 'last-entry' fields may be missing. Before this patch hedis
failed to read a reponse and filed with `Bulk Nothing` exception.

In this patch we provide a means to support that by introducing an
extra constructor to the XInfoStreamResponse. We introduced a
constructor instead of wrapping fields into Maybe, because this
change leads to a fewer refactoring and in addition in maintains
an invariant that the first and the last entries are either full or
empty simultaniously.